+++ backends/opengl/doc.go
@@ -1,5 +1,5 @@
-// Package pixelgl implements efficient OpenGL targets and utilities for the Pixel game development
+// Package opengl implements efficient OpenGL targets and utilities for the Pixel game development
 // library, specifically Window and Canvas.
 //
 // It also contains a few additional utilities to help extend Pixel with OpenGL graphical effects.
-package pixelgl
+package opengl